Display function in service, footprint_to_extent - in layer_utils
